requirements.txt
numpy>=1.19.0
pandas>=1.2.0
scipy>=1.7.0
psutil>=5.8.0
typing>=3.7.4
pyyaml>=6.0
numba>=0.53.0
cupy>=9.0.0
tensorflow>=2.5.0
torch>=1.9.0
jax>=0.2.0
dask>=2021.3.0
polars>=0.8.0
matplotlib>=3.4.0
pyarrow>=4.0.0